Crashes in WebCore::DocumentMarkerController::removeMarkersFromMarkerMapVectorPair() when deleting multiple lines of text. https://bugs.webkit.org/show_bug.cgi?id=48918 <rdar://problem/8620602>
Patch by Jia Pu <jpu@apple.com> on 2010-11-03
Reviewed by Darin Adler.

I haven't been able to found a reliable way to reproduce the bug. However, whenever it happens,
the crash is caused by a null node pointer returned by TextIterator. So it seems to be a
safe fix to guard against that.

* editing/Editor.cpp: (WebCore::Editor::removeSpellAndCorrectionMarkersFromWordsToBeEdited):



Identifier: 61776@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@71284 268f45cc-cd09-0410-ab3c-d52691b4dbfc