userscripts/user-script-*-document.html tests are flaky on Windows https://bugs.webkit.org/show_bug.cgi?id=45980
Reviewed by Eric Carlson.

We can't use counting.mp4 as our video file to create a video document, because on Windows counting.mp4
is opened in Windows Explorer or Quicktime instead of Safari. This bug is being tracked by
<https://bugs.webkit.org/show_bug.cgi?id=25220>.

This patch switches user-script-video-document.html to use silence-loop.mov instead of counting.mp4,
which creates a video document on both Mac and Windows.

* userscripts/resources/counting.mp4: Removed.
* userscripts/resources/silence-loop.mov: Added.
* userscripts/user-script-video-document-expected.txt:
* userscripts/user-script-video-document.html:



Identifier: 61920@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@71433 268f45cc-cd09-0410-ab3c-d52691b4dbfc