Pasting large amounts of plain text in a text area is very slow https://bugs.webkit.org/show_bug.cgi?id=49040 <rdar://problem/7640912>
Reviewed by Adele Peterson.

The optimization consists in skipping the ApplyStyle command when
pasting into a text area or into an element that has -webkit-user-modify: read-write-plaintext-only.
The biggest performance hit comes from iterating through
a large number of paragraph when no style change is needed.
A further step will be to improve the paragraph iteration when it
can't be avoided.

* editing/ReplaceSelectionCommand.cpp:
(WebCore::ReplaceSelectionCommand::doApply): Check if the selection
is in a text form control or in an element that has -webkit-user-modify: read-write-plaintext-only
and force no style matching.



Identifier: 61946@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@71459 268f45cc-cd09-0410-ab3c-d52691b4dbfc