<rdar://problem/8731320> WebProcess's CFNetwork cache is in ~/Library/Caches/com.apple.WebProcess
Reviewed by Sam Weinig.

* UIProcess/Launcher/mac/ProcessLauncherMac.mm:
(WebKit::ProcessLauncher::launchProcess): Get the location, disk capacity and memory capacity
of the UI process's NSURLCache and pass them to the WebProcess.
* WebProcess/mac/WebProcessMainMac.mm:
(WebKit::WebProcessMain): If an NSURLCache location was passed on the command line, set the
shared NSURLCache to that path and use the specified capacities as initial values.



Identifier: 64522@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@74175 268f45cc-cd09-0410-ab3c-d52691b4dbfc