[Qt] QML WebView does not transfer focus to underlying QGraphicsWebView
QML WebView creates QGraphicsWebView as it's child. The WebView is not
setup as a focus proxy of the QGraphicsWebView. Thus, even though the
WebView gets focus, the QGraphicsWebView does not get focus.
In QML, focus proxies or "focus scopes" are created using FocusScope.
This change makes WebView a FocusScope and sets the focus on the
QGraphicsWebView.

Reviewed by Simon Hausmann.
Acked by Aaron Kennedy <aaron.kennedy@nokia.com>

https://bugs.webkit.org/show_bug.cgi?id=51094

* declarative/qdeclarativewebview.cpp:
(QDeclarativeWebView::init):

Identifier: 65442@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@75119 268f45cc-cd09-0410-ab3c-d52691b4dbfc