Rolled back in r76078, with crash fixed. https://bugs.webkit.org/show_bug.cgi?id=52668
Reviewed by Darin Adler.

* runtime/JSGlobalObject.cpp:
(JSC::JSGlobalObject::markChildren): Account for the fact that the global
object moves its variables into and out of the register file. While out
of the register file, the symbol table's size is not an accurate count
for the size of the register array, since the BytecodeGenerator might
be compiling, adding items to the symbol table.



Identifier: 66431@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@76193 268f45cc-cd09-0410-ab3c-d52691b4dbfc