Don't create the Direct3D device before it's first needed
We only need the device once we decide to render. There's no point in creating it before
then.

Reviewed by Sam Weinig.

* platform/graphics/ca/win/CACFLayerTreeHost.cpp:
(WebCore::CACFLayerTreeHost::setWindow): Removed the call to createRenderer() from here.
We already have code to create it when we first try to draw.
(WebCore::CACFLayerTreeHost::createRenderer): Flush the context after we set our layer's
bounds so that the bounds will take effect the next time we render (which could be just
after this function returns).

Identifier: 66982@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@76764 268f45cc-cd09-0410-ab3c-d52691b4dbfc